perf: Lazy load web3inbox
